requirements.txt
# ---------------------------------------------------------------------------
# Web framework
# ---------------------------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# ---------------------------------------------------------------------------
# Database
# ---------------------------------------------------------------------------
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10  # used by Alembic sync runner

# ---------------------------------------------------------------------------
# Validation / settings
# ---------------------------------------------------------------------------
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic[email]==2.10.4

# ---------------------------------------------------------------------------
# Auth
# ---------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# ---------------------------------------------------------------------------
# Background tasks
# ---------------------------------------------------------------------------
celery[redis]==5.4.0
redis==5.2.1
flower==2.0.1
kombu==5.4.2

# ---------------------------------------------------------------------------
# Observability
# ---------------------------------------------------------------------------
structlog==24.4.0
sentry-sdk[fastapi,celery]==2.19.2

# ---------------------------------------------------------------------------
# HTTP client (for LLM API calls, webhooks, etc.)
# ---------------------------------------------------------------------------
httpx==0.28.1
